+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.2] - 2026-02-08
+### Added
+- Regex pattern can be configured
+- Scan opened editors if they weren't covered by the workspace scan on startup
+
+### Changed
+- Default regex match made a little more strict
+
 ## [0.0.1] - 2026-02-07
 ### Added
 - Initial release